Nixify - many changes
This is a large MR which enables this repository for nix. We were already using a nix flake, however this MR integrates a full environment as described in pyaion, with all packages built via mach-nix and requirements listed in a user friendly way. It also adds automatic documentation generation, units tests and pip vs nix dependency syncing, all running in the CI pipeline.